pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "overseerr-mcp"
version = "0.1.0"
description = "MCP server for Overseerr media request management"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.0.0",
    "pydantic>=2.0.0",
    "aiohttp>=3.8.0",
    "python-dotenv>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
overseerr-mcp = "src.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]
